fix build warnings

+++ sdk/storage/azure-storage-datamovement/src/job_engine_task_generation.cpp
@@ -160,7 +160,7 @@ namespace Azure { namespace Storage { namespace _detail {
         {
           auto task = m_rootTask->CreateTask<Blobs::_detail::UploadBlobFromFileTask>(
               _internal::TaskType::NetworkUpload, source, destination);
-          task->MemoryCost = taskModel.ObjectSize;
+          task->MemoryCost = static_cast<size_t>(taskModel.ObjectSize);
           task->JournalContext = JournalContext{jobPart, bitmapOffset};
           tasks.emplace_back(std::move(task));
           ++bitmapOffset;
@@ -192,7 +192,7 @@ namespace Azure { namespace Storage { namespace _detail {
             task->Context = context;
             task->BlockId = i;
             task->Offset = i * taskModel.ChunkSize;
-            task->Length = std::min<int64_t>(context->FileSize - task->Offset, taskModel.ChunkSize);
+            task->Length = std::min<uint64_t>(context->FileSize - task->Offset, taskModel.ChunkSize);
             task->MemoryCost = task->Length;
             task->JournalContext = JournalContext{jobPart, bitmapOffset};
             tasks.emplace_back(std::move(task));
@@ -241,7 +241,7 @@ namespace Azure { namespace Storage { namespace _detail {
               _internal::TaskType::NetworkDownload);
           task->Context = context;
           task->Offset = i * taskModel.ChunkSize;
-          task->Length = std::min<int64_t>(context->FileSize - task->Offset, taskModel.ChunkSize);
+          task->Length = std::min<uint64_t>(context->FileSize - task->Offset, taskModel.ChunkSize);
           task->MemoryCost = task->Length;
           task->JournalContext = JournalContext{jobPart, bitmapOffset};
           tasks.emplace_back(std::move(task));
